🔥 remove(tsconfig.json): will not use as root file
To not have ambiguity, we will not use the tsconfig.json as the root file.
We will use the tsconfig.base.json as the root file.

The next commit will be to rename the tsconfig.node.json to
tsconfig.base.json.

The reason to rename is only to make more sense to the name.